Read the properties-below toggle live, and make keyframe size fine and small
The toggle was read once when the clip opened, so flipping it in settings left an
open editor where it was. It is read every frame now through a supplier, and the
side/below layout is rebuilt when the answer changes, so it takes effect at once.

Keyframe size becomes a float capped at two pixels, in fine steps, instead of
whole pixels up to twelve - two was already larger than wanted, so the range is
where the useful sizes are and the steps are small enough to land on the one you
want. The shape offset is carried as a float end to end (interface and the six
renderers), the added amount is still the same for every shape, and the selected
highlight grows with it too.

+++ src/client/java/mchorse/bbs_mod/ui/framework/elements/input/keyframes/shapes/DiamondKeyframeShapeRenderer.java
@@ -23,7 +23,7 @@ public class DiamondKeyframeShapeRenderer implements IKeyframeShapeRenderer
     }
 
     @Override
-    public void renderKeyframe(UIContext uiContext, BufferBuilder builder, Matrix4f matrix, int x, int y, int offset, int c)
+    public void renderKeyframe(UIContext uiContext, BufferBuilder builder, Matrix4f matrix, int x, int y, float offset, int c)
     {
         float fOffset = offset * 1.5F;
