+++ src/client/java/mchorse/bbs_mod/ui/framework/elements/input/keyframes/shapes/SquareKeyframeShapeRenderer.java
@@ -23,8 +23,8 @@ public class SquareKeyframeShapeRenderer implements IKeyframeShapeRenderer
     }
 
     @Override
-    public void renderKeyframe(UIContext uiContext, BufferBuilder builder, Matrix4f matrix4f, int x, int y, int offset, int c)
+    public void renderKeyframe(UIContext uiContext, BufferBuilder builder, Matrix4f matrix4f, int x, int y, float offset, int c)
     {
-        uiContext.batcher.fillRect(builder, matrix4f, x - offset, y - offset, offset * 2, offset * 2, c, c, c, c);
+        uiContext.batcher.fillRect(builder, matrix4f, x - offset, y - offset, offset * 2F, offset * 2F, c, c, c, c);
     }
 }
